feat: show location cards where a location is given
