+++ components/custom-primary-details-actor.vue
@@ -13,32 +13,6 @@ interface Place {
 	relationType: RelationType | null
 }
 
-// TODO: Make use of first and last event if no places are available
-const places = computed(() => {
-	return props.entity.relations?.reduce((acc: Array<Place>, relation) => {
-		if(relation.relationSystemClass !== "object_location") return acc;
-		if(!relation.label || !relation.relationTo || !relation.relationType) return acc;
-		const id = getUnprefixedId(relation.relationTo);
-		const relationType = extractRelationTypeFromRelationString(relation.relationType);
-		const label = relation.label;
-		if(!id) return acc;
-		return [
-			...acc,
-			{
-				label,
-				id,
-				relationType
-			}]
-	}, []);
-});
-
-const getRelationTitle = (relation: RelationType) => {
-	if(props.entity.systemClass === 'person') {
-		return useRelationTitle(relation, 'person')
-	}
-	return useRelationTitle(relation)
-}
-
 const getRelationGroupTitle = (relation: RelationType) => {
 	if(props.entity.systemClass === 'person') {
 		return useRelationGroupTitle(relation, 'person')
@@ -104,26 +78,6 @@ onMounted(() => {
 </script>
 
 <template>
-	<div class="flex w-full flex-row flex-wrap gap-4">
-		<template
-			v-for="(place) in places"
-			:key="place.label || `place-${index}`"
-		>
-			<InfoCard
-				v-if="place.relationType"
-				class="max-w-48 p-4"
-				:icon="MapPinIcon"
-				:title="getRelationTitle(place.relationType)"
-				>
-				<template #content>
-					<EntityPreviewLink v-if="place.id" :id="useToNumber(place.id).value">
-						{{ place.label }}
-					</EntityPreviewLink>
-					{{ place.id ? '' : place.label }}
-				</template>
-			</InfoCard>
-		</template>
-	</div>
 	<GroupedRelationCollapsible
 		v-for="rel in collapsibleRelations"
 		:key="rel.relationType.crmCode + rel.relationType.inverse"
