+++ components/custom-primary-details-place.vue
@@ -1,18 +1,8 @@
 <script setup lang="ts">
 
-const {getUnprefixedId} = useIdPrefix();
-
 const t = useTranslations();
 
 const props = defineProps<{entity: EntityFeature}>();
-
-const getRelationTitle = (relation: RelationType) => {
-	return useRelationTitle(relation)
-}
-
-const getRelationGroupTitle = (relation: RelationType) => {
-	return useRelationGroupTitle(relation)
-}
 
 const collapsibleRelations: Array<{
 	relationType: RelationType,
@@ -54,13 +44,12 @@ onMounted(() => {
 </script>
 
 <template>
-
-<GroupedRelationCollapsible
-		v-for="rel in collapsibleRelations"
-		:key="rel.relationType.crmCode + rel.relationType.inverse"
-		:title="rel.title"
-		:relations="entity.relations"
-		:system-class="rel.systemClass"
-		:relation-type="rel.relationType"
-	/>
+	<GroupedRelationCollapsible
+			v-for="rel in collapsibleRelations"
+			:key="rel.relationType.crmCode + rel.relationType.inverse"
+			:title="rel.title"
+			:relations="entity.relations"
+			:system-class="rel.systemClass"
+			:relation-type="rel.relationType"
+		/>
 </template>
